requirements.txt
google-cloud-speech 
google-cloud-texttospeech 
transformers
torch
pydub
soundfile
Flask
ffmpeg-python